phy/lpddr4: extract common utilities
