+++ test/test_lpddr4.py
@@ -15,179 +15,45 @@ from typing import Mapping, Sequence
 from migen import *
 
 from litedram.phy import dfi
-from litedram.phy.lpddr4.simphy import Serializer, Deserializer
+from litedram.phy.utils import chunks, bit
 from litedram.phy.lpddr4.simphy import LPDDR4SimPHY, DoubleRateLPDDR4SimPHY
 from litedram.phy.lpddr4 import simsoc
 
-from litex.gen.sim import run_simulation as _run_simulation
+from test.test_phy_utils import run_simulation as _run_simulation
 
 
-def bit(n, val):
-    return (val & (1 << n)) >> n
+# Migen simulator supports reset signals so we could add CRG to start all the signals
+# in the same time, however the clock signals will still be visible in the VCD dump
+# and the generators we assign to them will still work before reset. For this reason we
+# use clocks set up in such a way that we have all the phase aligned clocks start in tick
+# 1 (not zero), so that we avoid any issues with clock alignment.
+#
+# NOTE: On hardware proper reset must be ensured!
+#
+# The simulation should start like this:
+#   sys          |_--------------
+#   sys_11_25    |___------------
+#   sys8x        |_----____----__
+#   sys8x_ddr    |_--__--__--__--
+#   sys8x_90     |___----____----
+#   sys8x_90_ddr |-__--__--__--__
+#
+# sys8x_90_ddr does not trigger at the simulation start (not an edge),
+# BUT a generator starts before first edge, so a `yield` is needed to wait until the first
+# rising edge!
+CLOCKS = {
+    "sys":          (64, 31),
+    "sys_11_25":    (64, 29),  # aligned to sys8x_90 (phase shift of 11.25)
+    "sys2x":        (32, 15),
+    "sys8x":        ( 8,  3),
+    "sys8x_ddr":    ( 4,  1),
+    "sys8x_90":     ( 8,  1),
+    "sys8x_90_ddr": ( 4,  3),
+}
 
-def chunks(lst, n):
-    for i in range(0, len(lst), n):
-        yield lst[i:i + n]
+def run_simulation(dut, generators, **kwargs):
+    _run_simulation(dut, generators, CLOCKS, **kwargs)
 
-
-def run_simulation(dut, generators, debug_clocks=False, **kwargs):
-    # Migen simulator supports reset signals so we could add CRG to start all the signals
-    # in the same time, however the clock signals will still be visible in the VCD dump
-    # and the generators we assign to them will still work before reset. For this reason we
-    # use clocks set up in such a way that we have all the phase aligned clocks start in tick
-    # 1 (not zero), so that we avoid any issues with clock alignment.
-    #
-    # NOTE: On hardware proper reset must be ensured!
-    #
-    # The simulation should start like this:
-    #   sys          |_--------------
-    #   sys_11_25    |___------------
-    #   sys8x        |_----____----__
-    #   sys8x_ddr    |_--__--__--__--
-    #   sys8x_90     |___----____----
-    #   sys8x_90_ddr |-__--__--__--__
-    #
-    # sys8x_90_ddr does not trigger at the simulation start (not an edge),
-    # BUT a generator starts before first edge, so a `yield` is needed to wait until the first
-    # rising edge!
-    clocks = {
-        "sys":          (64, 31),
-        "sys_11_25":    (64, 29),  # aligned to sys8x_90 (phase shift of 11.25)
-        "sys2x":        (32, 15),
-        "sys8x":        ( 8,  3),
-        "sys8x_ddr":    ( 4,  1),
-        "sys8x_90":     ( 8,  1),
-        "sys8x_90_ddr": ( 4,  3),
-    }
-
-    if debug_clocks:
-        class DUT(Module):
-            def __init__(self, dut):
-                self.submodules.dut = dut
-                for clk in clocks:
-                    setattr(self.clock_domains, "cd_{}".format(clk), ClockDomain(clk))
-                    cd = getattr(self, 'cd_{}'.format(clk))
-                    self.comb += cd.rst.eq(0)
-
-                    s = Signal(4, name='dbg_{}'.format(clk))
-                    sd = getattr(self.sync, clk)
-                    sd += s.eq(s + 1)
-        dut = DUT(dut)
-
-    _run_simulation(dut, generators, clocks, **kwargs)
-
-
-class TestSimSerializers(unittest.TestCase):
-    @staticmethod
-    def data_generator(i, datas):
-        for data in datas:
-            yield i.eq(data)
-            yield
-        yield i.eq(0)
-        yield
-
-    @staticmethod
-    def data_checker(o, datas, n, latency, yield1=False):
-        if yield1:
-            yield
-        for _ in range(latency):
-            yield
-        for _ in range(n):
-            datas.append((yield o))
-            yield
-        yield
-
-    def serializer_test(self, *, data_width, datas, clk, clkdiv, latency, clkgen=None, clkcheck=None, **kwargs):
-        clkgen = clkgen if clkgen is not None else clkdiv
-        clkcheck = clkcheck if clkcheck is not None else clk
-
-        received = []
-        dut = Serializer(clk=clk, clkdiv=clkdiv, i_dw=data_width, o_dw=1)
-        generators = {
-            clkgen: self.data_generator(dut.i, datas),
-            clkcheck: self.data_checker(dut.o, received, n=len(datas) * data_width, latency=latency * data_width, yield1=True),
-        }
-        run_simulation(dut, generators, **kwargs)
-
-        received = list(chunks(received, data_width))
-        datas  = [[bit(i, d) for i in range(data_width)] for d in datas]
-        self.assertEqual(received, datas)
-
-    def deserializer_test(self, *, data_width, datas, clk, clkdiv, latency, clkgen=None, clkcheck=None, **kwargs):
-        clkgen = clkgen if clkgen is not None else clk
-        clkcheck = clkcheck if clkcheck is not None else clkdiv
-
-        datas = [[bit(i, d) for i in range(data_width)] for d in datas]
-
-        received = []
-        dut = Deserializer(clk=clk, clkdiv=clkdiv, i_dw=1, o_dw=data_width)
-        generators = {
-            clkgen: self.data_generator(dut.i, itertools.chain(*datas)),
-            clkcheck: self.data_checker(dut.o, received, n=len(datas), latency=latency),
-        }
-
-        run_simulation(dut, generators, **kwargs)
-
-        received = [[bit(i, d) for i in range(data_width)] for d in received]
-        self.assertEqual(received, datas)
-
-    DATA_8 = [0b11001100, 0b11001100, 0b00110011, 0b00110011, 0b10101010]
-    DATA_16 = [0b1100110011001100, 0b0011001100110011, 0b0101010101010101]
-
-    ARGS_8 = dict(
-        data_width = 8,
-        datas = DATA_8,
-        clkdiv = "sys",
-        clk = "sys8x",
-        latency = Serializer.LATENCY,
-    )
-
-    ARGS_16 = dict(
-        data_width = 16,
-        datas = DATA_16,
-        clkdiv = "sys",
-        clk = "sys8x_ddr",
-        latency = Serializer.LATENCY,
-    )
-
-    def _s(default, **kwargs):
-        def test(self):
-            new = default.copy()
-            new.update(kwargs)
-            self.serializer_test(**new)
-        return test
-
-    def _d(default, **kwargs):
-        def test(self):
-            new = default.copy()
-            new["latency"] = Deserializer.LATENCY
-            new.update(kwargs)
-            self.deserializer_test(**new)
-        return test
-
-    test_sim_serializer_8 = _s(ARGS_8)
-    test_sim_serializer_8_phase90 = _s(ARGS_8, clkdiv="sys_11_25", clk="sys8x_90")
-    # when clkgen and clkdiv are not phase aligned (clkdiv is delayed), there will be lower latency
-    test_sim_serializer_8_phase90_gen0 = _s(ARGS_8, clkdiv="sys_11_25", clk="sys8x_90", clkgen="sys", latency=Serializer.LATENCY - 1)
-    test_sim_serializer_8_phase90_check0 = _s(ARGS_8, clkdiv="sys_11_25", clk="sys8x_90", clkcheck="sys8x")
-
-    test_sim_serializer_16 = _s(ARGS_16)
-    test_sim_serializer_16_phase90 = _s(ARGS_16, clkdiv="sys_11_25", clk="sys8x_90_ddr")
-    test_sim_serializer_16_phase90_gen0 = _s(ARGS_16, clkdiv="sys_11_25", clk="sys8x_90_ddr", clkgen="sys", latency=Serializer.LATENCY - 1)
-    test_sim_serializer_16_phase90_check0 = _s(ARGS_16, clkdiv="sys_11_25", clk="sys8x_90_ddr", clkcheck="sys8x_ddr")
-
-    # for phase aligned clocks the latency will be bigger (preferably avoid phase aligned reading?)
-    test_sim_deserializer_8 = _d(ARGS_8, latency=Deserializer.LATENCY + 1)
-    test_sim_deserializer_8_check90 = _d(ARGS_8, clkcheck="sys_11_25")
-    test_sim_deserializer_8_gen90_check90 = _d(ARGS_8, clkcheck="sys_11_25", clkgen="sys8x_90")
-    test_sim_deserializer_8_phase90 = _d(ARGS_8, clkdiv="sys_11_25", clk="sys8x_90", latency=Deserializer.LATENCY + 1)
-    test_sim_deserializer_8_phase90_check0 = _d(ARGS_8, clkdiv="sys_11_25", clk="sys8x_90", clkcheck="sys", latency=Deserializer.LATENCY + 1)
-
-    test_sim_deserializer_16 = _d(ARGS_16, latency=Deserializer.LATENCY + 1)
-    test_sim_deserializer_16_check90 = _d(ARGS_16, clkcheck="sys_11_25")
-    test_sim_deserializer_16_gen90_check90 = _d(ARGS_16, clkcheck="sys_11_25", clkgen="sys8x_90_ddr")
-    test_sim_deserializer_16_phase90 = _d(ARGS_16, clkdiv="sys_11_25", clk="sys8x_90_ddr", latency=Deserializer.LATENCY + 1)
-    test_sim_deserializer_16_phase90_check0 = _d(ARGS_16, clkdiv="sys_11_25", clk="sys8x_90_ddr", clkcheck="sys", latency=Deserializer.LATENCY + 1)
 
 
 BOLD = '\033[1m'
